style(tooltip): unarchive button; refactor: `useArchiveHandler` and `ArchiveButton`
