+++ client/src/components/Conversations/ConvoOptions/index.ts
@@ -1,4 +1,3 @@
-export { default as ArchiveButton } from './ArchiveButton';
 export { default as DeleteButton } from './DeleteButton';
 export { default as ShareButton } from './ShareButton';
 export { default as SharedLinkButton } from './SharedLinkButton';
